requirements.txt
APScheduler==3.9.1
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
CacheControl==0.12.10
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==37.0.2
DateTime==4.4
Deprecated==1.2.13
firebase-admin==5.2.0
Flask==2.1.1
gcloud==0.17.0
google-api-core==2.7.1
google-api-python-client==2.43.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
google-cloud-core==2.2.3
google-cloud-firestore==2.4.0
google-cloud-storage==2.2.1
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-status==1.44.0
h11==0.13.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
jsonify==0.5
jws==0.1.3
MarkupSafe==2.1.1
msgpack==1.0.3
numpy==1.22.3
oauth2client==3.0.0
outcome==1.1.0
pandas==1.4.2
proto-plus==1.20.3
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.4.3
pykrx==1.0.32
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
python-dateutil==2.8.2
python-jwt==2.0.1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-toolbelt==0.7.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.20.0
trio-websocket==0.9.2
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.1.1
wrapt==1.14.0
wsproto==1.1.0
xlrd==2.0.1
zipp==3.8.0
zope.interface==5.4.0
